Update and rename terraform-project3 to terraform+aws-project-3
Project 3: This Terraform configuration automates the creation of a VPC with both public and private subnets, an internet gateway, and a public route table, along with the necessary security groups to allow HTTP, HTTPS, and SSH traffic. It also provisions an EC2 instance in the us-east-1 region, configured with a dynamic and customizable Python Flask app, which is uploaded to the instance via a file provisioner and executed using a remote-exec provisioner. All settings such as the VPC CIDR, instance type, AMI, and SSH key paths are defined as variables, allowing for flexible and reusable infrastructure setup.